# Optimize the ST_XMin, ST_XMax, ST_YMin, ST_YMax functions