Similar to [#9953](https://github.com/prestodb/presto/pull/9953), ST_XMin, ST_XMax, ST_YMin, and ST_YMax geospatial functions spend most of the time deserializing geometry from Slice. This PR updates them to limit deserialization to just the envelope.

This change fixes the NullPointerException reported in [#10027](https://github.com/prestodb/presto/pull/10027) as well.

Micro-benchmarking shows a big performance improvement.

Before:

```
Benchmark                       Mode  Cnt Score     Error    Units
BenchmarkSTXmin.complexGeometry avgt 20 123830.375 ± 4063.732 ns/op
BenchmarkSTXmin.simpleGeometry avgt 20 535.399 ± 56.854 ns/op`
```

and after:

```
Benchmark                       Mode  Cnt Score   Error  Units
BenchmarkSTXmin.complexGeometry avgt 20 59.393 ± 5.843 ns/op
BenchmarkSTXmin.simpleGeometry avgt 20 54.984 ± 1.887 ns/op
```