@mehrdad-honarkhah 2018-02-28
Thanks @nezihyigitbasi I made those functions static and made sure the lines are wrapped at exactly 72 characters width.