@mbasmanova 2018-03-01
@mehrdad-honarkhah Mehrdad, I'm seeing build failures due to checkstyle errors:

https://api.travis-ci.org/v3/job/347538801/log.txt

```
[INFO] --- maven-checkstyle-plugin:2.17:check (checkstyle) @ presto-geospatial ---
[INFO] There are 3 errors reported by Checkstyle 8.7 with checkstyle/airbase-checks.xml ruleset.
[ERROR] src/test/java/com/facebook/presto/plugin/geospatial/BenchmarkSTXMin.java:[17,8] (imports) UnusedImports: Unused import - io.airlift.slice.Slices.
[ERROR] src/test/java/com/facebook/presto/plugin/geospatial/BenchmarkSTXMin.java:[39] (imports) ImportOrder: Wrong order for 'com.facebook.presto.plugin.geospatial.GeoFunctions.stXMin' import.
[ERROR] src/test/java/com/facebook/presto/plugin/geospatial/BenchmarkSTXMin.java:[40] (imports) ImportOrder: Wrong order for 'com.facebook.presto.plugin.geospatial.GeoFunctions.stGeometryFromText' import.
```
Would you take a look and fix these?