@mehrdad-honarkhah 2018-03-01
@mbasmanova Thanks. I have made the necessary changes for style checks and pushed it.