@mehrdad-honarkhah 2018-03-01
My previous push did not include my changes. So I changed the following in this push:
(1) added null tests
(2) changed the commit message
(3) stylechecks on imports are now going through in this commit.